Sort patches to new directories
