+++ Fika.Dedicated/FikaDedicatedPlugin.cs
@@ -16,6 +16,9 @@ using Fika.Core.UI.Patches;
 using Fika.Dedicated.Classes;
 using Fika.Dedicated.Patches;
 using Fika.Dedicated.Patches.DestroyGraphics;
+using Fika.Dedicated.Patches.DLSS;
+using Fika.Dedicated.Patches.TextureValidateFormat;
+using Fika.Dedicated.Patches.VRAM;
 using HarmonyLib;
 using Newtonsoft.Json;
 using SPT.Common.Http;
