+++ Fika.Dedicated/Patches/DestroyGraphics/AmbientLight_Start_Patch.cs
@@ -2,7 +2,7 @@
 using System.Reflection;
 using UnityEngine;
 
-namespace Fika.Dedicated.Patches
+namespace Fika.Dedicated.Patches.DestroyGraphics
 {
 	internal class AmbientLight_Start_Patch : ModulePatch
 	{
@@ -14,7 +14,7 @@ namespace Fika.Dedicated.Patches
 		[PatchPrefix]
 		public static bool Prefix(AmbientLight __instance)
 		{
-			GameObject.Destroy(__instance);
+			Object.Destroy(__instance);
 			return false;
 		}
 	}
