+++ Fika.Dedicated/Patches/DestroyGraphics/TextureDecalsPainter_Awake_Patch.cs
@@ -2,7 +2,7 @@
 using System.Reflection;
 using UnityEngine;
 
-namespace Fika.Dedicated.Patches
+namespace Fika.Dedicated.Patches.DestroyGraphics
 {
 	internal class TextureDecalsPainter_Awake_Patch : ModulePatch
 	{
@@ -14,7 +14,7 @@ namespace Fika.Dedicated.Patches
 		[PatchPrefix]
 		public static bool Prefix(TextureDecalsPainter __instance)
 		{
-			GameObject.Destroy(__instance);
+			Object.Destroy(__instance);
 			return false;
 		}
 	}
