+++ Fika.Dedicated/Patches/VRAM/VRAMPatch3.cs
@@ -1,7 +1,7 @@
 ﻿using SPT.Reflection.Patching;
 using System.Reflection;
 
-namespace Fika.Dedicated.Patches
+namespace Fika.Dedicated.Patches.VRAM
 {
 	// Token: 0x0200000B RID: 11
 	public class VRAMPatch3 : ModulePatch
