+++ Fika.Dedicated/Patches/VRAM/VRAMPatch4.cs
@@ -2,7 +2,7 @@
 using SPT.Reflection.Patching;
 using System.Reflection;
 
-namespace Fika.Dedicated.Patches
+namespace Fika.Dedicated.Patches.VRAM
 {
 	// Token: 0x0200000C RID: 12
 	public class VRAMPatch4 : ModulePatch
